Move QuestionnaireViewModelTest in eir to engine module
